fix: 3 pipeline bugs causing runtime NameError and TypeError
- 05_annotate.py: annotation_confidence dtype was Categorical,
  preventing numeric comparison (< 0.02). Added .astype(float).
- 06_exploratory.py: plot_composition referenced CFG.table_dir
  without CFG in scope. Added table_dir parameter.
- 08_trajectory.py: compute_dpt referenced CFG.n_diffmap_comps
  without CFG in scope. Added CFG parameter.

+++ scripts/05_annotate.py
@@ -60,7 +60,7 @@ def run_annotation(adata, marker_dict, log):
             lambda row: row.sort_values(ascending=False).values, axis=1, result_type='expand'
         )
         confidence = sorted_scores.iloc[:, 0] - sorted_scores.iloc[:, 1]
-        adata.obs['annotation_confidence'] = adata.obs['leiden'].map(confidence).values
+        adata.obs['annotation_confidence'] = adata.obs['leiden'].map(confidence).astype(float).values
         low_conf = (adata.obs['annotation_confidence'] < 0.02).sum()
         if low_conf > 0:
             log.info("  低置信度细胞 (<0.02): %d (%.1f%%)",
